Move components/image.js to controls/image.js
components/image.js was just a wrapper
that used <Image>,
so move that wrapper into the component.
