+++ js/src/block-editor/components/index.js
@@ -1,5 +1,4 @@
 export { default as GcbInspector } from './gcb-inspector';
 export { default as Edit } from './edit';
 export { default as Fields } from './fields';
-export { default as Image } from './image';
 export { default as RepeaterRows } from './repeater-rows';
